+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-05-28
+
+### Added
+- Matching single and array translators
+- Import for correctly namespaced exception
+
+### Fixed
+- Wrong namespace configuration for exception
+
 ## 1.0.1 - 2020-04-12
 
 ### Changed
@@ -15,5 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created the initial implementation of the translator package.
 
 # Versions
-- [1.0.1 > Unreleased](https://github.com/grizz-it/translator/compare/1.0.1...HEAD)
+- [1.1.0 > Unreleased](https://github.com/grizz-it/translator/compare/1.1.0...HEAD)
+- [1.0.1 > 1.1.0](https://github.com/grizz-it/translator/compare/1.0.1...1.1.0)
 - [1.0.0 > 1.0.1](https://github.com/grizz-it/translator/compare/1.0.0...1.0.1)